requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
docx2pdf==0.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.2
MarkupSafe==3.0.2
numpy==2.2.4
pandas==2.2.3
pdf2image==1.17.0
pillow==11.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
pywin32==310
six==1.17.0
SQLAlchemy==2.0.40
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
